requirements.txt
flask
flask-sqlalchemy
flask-login
flask-wtf
flask-socketio
redis
python-dotenv